ME-721: Migrated AutomatedEnum.
