+++ .tmp/roomlistwatcher/common/utility.py
@@ -3,17 +3,6 @@
 import collections
 import time
 
-import enum
-
-
-class AutomatedEnum(enum.Enum):
-
-    def __new__(cls):
-        value = len(cls.__members__) + 1
-        object_ = object.__new__(cls)
-        object_._value_ = value
-        return object_
-
 
 class CountdownTimer(object):
